Feat #21 - 기능요청 수정,삭제 API 진행도가 Pending일때만 가능하게 변경

+++ src/featureRequest/featureRequests.service.ts
@@ -11,6 +11,7 @@ import { FeatureRequestTagRepository } from './repositiories/featureRequestTag.r
 import { EditFeatureRequestDto } from './dtos/edit-featureRequest.dto';
 import { GetFeatureRequestsDto } from './dtos/get-featureRequests.dto';
 import { User } from 'src/users/entities/user.entity';
+import { Progress } from './types/progressType';
 @Injectable()
 export class FeatureRequestService {
   constructor(
@@ -95,6 +96,12 @@ export class FeatureRequestService {
     if (featureRequest.authorId !== user.id) {
       throw new BadRequestException('작성자만 수정이 가능합니다');
     }
+    if (!this.canUpdateAndDelete(featureRequest.progress)) {
+      throw new BadRequestException(
+        '답변이 채택된 게시글은 수정이 불가능합니다.',
+      );
+    }
+    console.log(editFeatureRequest);
     await this.featureRequestRepository.save([
       { id: featureRequestId, ...editFeatureRequest },
     ]);
@@ -119,7 +126,16 @@ export class FeatureRequestService {
     if (featureRequest.authorId !== user.id) {
       throw new BadRequestException('작성자만 삭제가 가능합니다');
     }
+    if (!this.canUpdateAndDelete(featureRequest.progress)) {
+      throw new BadRequestException(
+        '답변이 채택된 게시글은 삭제가 불가능합니다.',
+      );
+    }
     await this.featureRequestRepository.delete({ id: featureRequest.id });
     return true;
+  }
+
+  canUpdateAndDelete(progress: Progress) {
+    return progress == Progress.Pending;
   }
 }
